Re-organization of components
