+++ src/components/Room/JoinRequests/index.js
@@ -1,0 +1,2 @@
+import JoinRequests from './JoinRequests';
+export default JoinRequests;
