+++ src/components/Room/Room.css
@@ -3,9 +3,8 @@
 }
 
 .room-container {
-    padding: 40px;
-    width: 80vw;
-    max-width: 400px;
+    height: 100vh;
+    width: 100vw;
 }
 
 .room-form {
