+++ src/components/Room/RoomBottomPanel/index.js
@@ -1,0 +1,2 @@
+import RoomBottomPanel from './RoomBottomPanel';
+export default RoomBottomPanel;
