+++ src/components/Room/RoomCanvas/index.js
@@ -1,0 +1,2 @@
+import RoomCanvas from './RoomCanvas';
+export default RoomCanvas;
