+++ src/components/Room/VideoGrid/VideoGrid.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import Video from '../../Video';
+
+function VideoGrid() {
+  // TODO: multiple video components to populate the grid
+  return (
+    <div>
+      <Video />
+    </div>
+  )
+}
+
+export default VideoGrid;
