+++ src/components/Room/VideoGrid/index.js
@@ -1,0 +1,2 @@
+import VideoGrid from './VideoGrid';
+export default VideoGrid;
