+++ src/components/Video/Video.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Video() {
+  // TODO: this is a placeholder for each "video" component. To be used for Vestibule media preview and Room's Video grid
+
+  return (
+    <div></div>
+  )
+}
+
+export default Video;
