+++ src/components/Video/index.js
@@ -1,0 +1,2 @@
+import Video from './Video';
+export default Video;
